Precedence of conflicting page processor context.
The order of execution of page processors was reversed in #315 so that custom page processors returning an HttpResponse would bypass the default processors.

That had the side-effect of making context variables in default processors overwrite those in custom processors, which isn't very intuitive. This change restores the original behaviour of context variables, while retaining the reversed execution order.